refactor(engine): split storage.ts into focused storage/ modules
Splits the 910-line storage service into 6 focused modules:
- storage/types.ts (interfaces, StorageError, SyncResult/Options)
- storage/config.ts (allowed types, cache control, validation, utils)
- storage/operations.ts (uploadFile, getFile, deleteFile, etc.)
- storage/metadata.ts (D1 queries: getFileRecord, listFiles, etc.)
- storage/sync.ts (R2 → D1 sync: syncFromBucket)
- storage/response.ts (shouldReturn304, buildFileHeaders)

Original storage.ts becomes a thin re-export barrel — all existing
import paths continue to work unchanged.
